% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build.r
\name{build_ssm}
\alias{build_ssm}
\title{Build model in SSM}
\usage{
build_ssm(model, pop_name, data, start_date, priors, inputs, reactions,
  observations, remainder = NULL, pop_size = NULL, diffed = NULL)
}
\arguments{
\item{model}{character, full path to model directory.}

\item{pop_name}{character, name of the population.}

\item{data}{data.frame, must have a column named \code{date} that contains
the observation dates (in \code{YYYY-MM-DD} format); and one or more numeric column(s)
that contains the observed states and which are named according to the convention
\code{x_obs}, where \code{x} is the name of the oberved state in the model. Missing
data are represented by \code{NA}.}

\item{start_date}{character, starting date of model integration (in \code{YYYY-MM-DD} format).}

\item{priors}{list of priors.}

\item{inputs}{list of inputs.}

\item{reactions}{list of reactions.}

\item{observations}{list of observations.}

\item{remainder}{character, name of the state variable considered as a remainder.}

\item{pop_size}{character, name of the parameter corresponding to the population size. Required if \code{remainder} is provided.}

\item{diffed}{character, parameter names that follow a diffusion}
}
\description{
This function build and compile the model in SSM.
}

