% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{calibrate_smc}
\alias{calibrate_smc}
\title{Calibrate the number of particles for a SMC}
\usage{
calibrate_smc(ssm, n_parts, n_replicates, plot = TRUE, ...)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{n_parts}{numeric vector containing the number of particles to test.}

\item{n_replicates}{for each value of \code{n_parts}, the number of replicates. The higher the more accurate the estimatation of the variance but it will take longer to compute.}

\item{plot}{logical, if \code{TRUE} (default) the function will print 2 diagnostic plots.}

\item{...}{other parameters to be passed to \code{\link{SMC}}.}
}
\value{
a list of 3 elements:
\itemize{
\item \code{smc} a \code{data_frame} containing one \code{ssm} object per SMC run.
\item \code{summary} a \code{data_frame} containing the SSM summary of each SMC run (loglikelihood etc.).
\item \code{plot} a list of 2 plots: \code{boxplot} and \code{mean_var}
}
}
\description{
This function allows you to optimize the number of particles in your SMC. It displays how the variance and the mean of the likelihood estimator change with increasing number of particles. The idea is then to select an intermediate number of particles such that the estimator is stable, thus saving further computation burden in your PMCMC.
}
\note{
This function will always run an SMC with PSR approximation since it's the only situation where you will be interested in calibrating the number of particles.
}
