% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap.r
\name{call_ssm}
\alias{call_ssm}
\title{Call to SSM}
\usage{
call_ssm(ssm, approx = c("ode", "sde", "psr"), do = c("kalman", "kmcmc",
  "ksimplex", "mif", "pmcmc", "simplex", "simul", "smc"),
  env = parent.frame())
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{approx}{character, approximation used to simulate \code{ssm}:
\itemize{
    \item \code{"ode"} deterministic approximation, based on Ordinary Differential Equations.
    \item \code{"sde"} diffusion approximation, based on Stochastic Differential Equations.
    \item \code{"psr"} Euler-multinomial approximation, based on Poisson process with Stochastic Rates.
}}

\item{do}{character, algorithm to perform on \code{ssm}}

\item{env}{calling environment}
}
\value{
a \code{ssm} object updated with latest SSM output and ready to be piped into another SSM block.
}
\description{
This is the core function to call SSM from R
}

