% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.r
\name{convert_symbol}
\alias{convert_symbol}
\title{Convert symbols for latex and html}
\usage{
convert_symbol(x, to = c("html", "latex"))
}
\arguments{
\item{x}{character vector containing symbols to be converted in latex or html.}

\item{to}{charcater, either \code{"html"} or \code{"latex"}}
}
\description{
Symbols must be expressed as a string to be converted.
}
