% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs.r
\name{do_lhs}
\alias{do_lhs}
\title{Perform LHS}
\usage{
do_lhs(
  ssm,
  n,
  do = c("kalman", "kmcmc", "ksimplex", "mif", "pmcmc", "simplex", "simul", "smc"),
  ...
)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{n}{numeric, sample size}

\item{do}{character, algorithm to perform on \code{ssm}}

\item{...}{additional arguments to be passed to \code{do}}
}
\value{
a \code{tbl} containing all \code{ssm} outputed by \code{do}.
}
\description{
Function to perform any SSM block on a LHS. The LHS is currently a random sample from the prior distribution.
}
\seealso{
\code{\link{get_max_lhs}}
}
