% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SSMinR-package.r
\name{ebola_2014}
\alias{ebola_2014}
\title{Time-series of the 2014 Ebola epidemic in Liberia}
\format{Data frames}
\source{
\url{http://apps.who.int/gho/data/view.ebola-sitrep.ebola-country-LBR-latest-download?embed=1}
}
\description{
Two datasets \code{liberia1} and \code{liberia2} containing the weekly incidence of confirmed and probable Ebola virus disease (EVD)
cases recorded during the 2014 Ebola outbreak in Liberia. The time-series span the period 2014-07-01 to 2014-08-25 for \code{liberia1}
and 2014-12-03 for \code{liberia2}.
}
\details{
\itemize{
  \item \code{date} calendar date of the record
  \item \code{incidence_obs} weekly count incidence of confirmed and probable EVD cases
}
}

