% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_hat}
\alias{get_hat}
\title{Compute hat for X}
\usage{
get_hat(df_X, hat)
}
\arguments{
\item{df_X}{A dataframe with 4 columns: date, index, variable and value.}

\item{hat}{A numeric vector containing the quantiles to compute, centered on the median. For example \code{hat = c(0.5, 0.95)} will compute the interquartile range as well as the 95% credible interval.}
}
\value{
A dataframe
}
\description{
Compute quantile intervals (hat) for states trajectories $X_t$.
}
