% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lhs.r
\name{get_max_lhs}
\alias{get_max_lhs}
\title{Get max LHS}
\usage{
get_max_lhs(lhs)
}
\arguments{
\item{lhs}{a \code{tbl}, as returned by \code{\link{do_lhs}}}
}
\value{
\code{ssm} object
}
\description{
Extract the \code{ssm} object with highest log-likelihood from \code{lhs}.
}
\seealso{
\code{\link{do_lhs}}
}

