% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrap.r
\name{ksimplex}
\alias{ksimplex}
\title{Run a Kalman-Simplex}
\usage{
ksimplex(ssm, dt = NULL, id = 0, root = NULL, iter = NULL,
  n_obs = NULL, size = NULL, eps_abs_integ = NULL, eps_rel_integ = NULL,
  like_min = NULL, freeze_forcing = NULL, interpolator = NULL,
  verbose = FALSE, warning = FALSE, no_dem_sto = FALSE,
  no_white_noise = FALSE, no_diff = FALSE, trace = TRUE, diag = TRUE,
  prior = FALSE, seed_time = TRUE)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{dt}{numeric, integration time step.}

\item{id}{integer, unique integer identifier that will be appended to the output.}

\item{root}{character, root path for output files (if any) (no trailing slash). If \code{NULL} (default), outputs are written in "your_model_path/the_name_of_the_wrapper".}

\item{iter}{numeric, number of iterations.}

\item{n_obs}{numeric, number of observations to be fitted (for tempering). If \code{NULL} (default), all observations are fitted.}

\item{size}{numeric, simplex size used as stopping criteria}

\item{eps_abs_integ}{numeric, absolute error for adaptive step-size control.}

\item{eps_rel_integ}{numeric, relative error for adaptive step-size control.}

\item{like_min}{numeric, particles with likelihood smaller than \code{like_min} are considered lost. If \code{NULL} (default) lower bound on likelihood based on machine precision.}

\item{freeze_forcing}{character, freeze covariates to their value at specified date (in \code{YYYY-MM-DD} format).}

\item{interpolator}{character, gsl interpolator for covariates}

\item{verbose}{logical, print logs (verbose). Default to \code{FALSE}.}

\item{warning}{logical, print warnings. Default to \code{FALSE}.}

\item{no_dem_sto}{logical, turn off demographic stochasticity (if any). Default to \code{FALSE}.}

\item{no_white_noise}{logical, turn off white noises (if any). Default to \code{FALSE}.}

\item{no_diff}{logical, turn off diffusions (if any). Default to \code{FALSE}.}

\item{trace}{logical, print the trace. Default to \code{TRUE}.}

\item{diag}{logical, print the diagnostics outputs (e.g. prediction residuals). Default to \code{TRUE}.}

\item{prior}{logical, add log(prior) to the estimated log-likelihood. Default to \code{TRUE}.}

\item{seed_time}{logical, seed the random number generator with the current time. Default to \code{TRUE}.}
}
\value{
a \code{ssm} object updated with latest SSM output and ready to be piped into another SSM block.
}
\description{
Function to run a Simplex using a Kalman filter to evaluate the log-likelihood. Only for the \code{sde} approximation of a \code{ssm}.
}

