% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMinR-package.r
\name{liberia1}
\alias{liberia1}
\title{Time-series of the Ebola epidemic in Liberia until August 2014}
\format{Data frames}
\source{
\url{http://apps.who.int/gho/data/view.ebola-sitrep.ebola-country-LBR-latest-download?embed=1}
}
\description{
This dataset contains the weekly incidence of confirmed and probable Ebola virus disease (EVD)
cases recorded during the 2014 Ebola outbreak in Liberia. The time-series span the period 2014-07-01 to 2014-08-25.
}
\details{
\itemize{
  \item \code{date} calendar date of the record
  \item \code{incidence_obs} weekly count incidence of confirmed and probable EVD cases
}
}

