% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.r
\name{obs}
\alias{obs}
\alias{poisson_obs}
\alias{discretized_normal_obs}
\alias{binomial_obs}
\title{Observation processes for SSM}
\usage{
poisson_obs(obs = NULL, state = NULL, reporting)

discretized_normal_obs(obs = NULL, state = NULL, reporting, overdispersion)

binomial_obs(x_obs = NULL, n_obs = NULL, p = NULL)
}
\arguments{
\item{obs}{character, name of the observation in the data}

\item{state}{character, name of the observed state in the model}

\item{reporting}{character, name of the reporting rate parameter}

\item{overdispersion}{character, name of the overdispersion parameter}

\item{x_obs}{character, name of the column with the number of successes in the data}

\item{n_obs}{character, name of the column with the sample size in the data}

\item{p}{character, probability of success, usually a combination of the states and parameters of the model (e.g. \code{p="I/N"})}
}
\description{
SSM currently accept the following observation processes.
}
