% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prior.r
\name{one_ssm_sample_prior}
\alias{one_ssm_sample_prior}
\title{Sample prior}
\usage{
one_ssm_sample_prior(ssm)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}
}
\value{
a \code{ssm} object
}
\description{
Generate one sample \code{theta} of \code{ssm} from the prior distribution
}
\seealso{
\code{\link{one_theta_sample_prior}}
}

