% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.r
\name{one_theta_sample_prior}
\alias{one_theta_sample_prior}
\title{Sample from prior}
\usage{
one_theta_sample_prior(priors, theta_names = NULL)
}
\arguments{
\item{priors}{a list of priors}

\item{theta_names}{character, names of the parameters sampled. If \code{NULL} (default) all parameters are sampled.}
}
\description{
Generate a sample from prior density distribution.
}

