% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_data}
\alias{plot_data}
\title{Plot data}
\usage{
plot_data(ssm, scales = "free_y")
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{scales}{character, should scales be \code{"fixed"}, \code{"free"}, or free in one dimension: \code{"free_x"}, \code{"free_y"} (the default).}
}
\description{
Plot the data of your \code{ssm} object.
}

