% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_model}
\alias{plot_model}
\title{Interactive display of SSM model}
\usage{
plot_model(ssm, collapse_erlang = TRUE)
}
\arguments{
\item{ssm}{a \code{\link{ssm}} object}

\item{collapse_erlang}{logical, if \code{TRUE} erlang compartments are collapsed into a single compartment to improve visibility.}
}
\description{
This function displays the SSM model as a force network in a web browser, allowing for manipulation.
}

