% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_priors}
\alias{plot_priors}
\title{Plot priors}
\usage{
plot_priors(ssm, theta_names = NULL, quantile_limits = c(0.01, 0.99),
  x_limits = NULL, plot = TRUE)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{theta_names}{character, specify which theta priors to plot. By default (=\colde{NULL}) all theta priors are plotted.}

\item{quantile_limits}{numeric, vector of length 2 specifying the limits of the priors in terms of quantile. Default to the 1% and 99% quantiles (=\code{c(0.01, 0.99)}).}

\item{x_limits}{numeric, vector of length 2 specifying the limits of the priors in terms of theta. Default to \code{NULL}.}

\item{plot}{logical, if \code{FALSE} a dataframe of prior values will be returned insread iof a plot}
}
\value{
either a \code{ggplot} or a \code{data.frame}
}
\description{
Plot prior distribution as specified in \code{ssm} object.
}
