% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_theta}
\alias{plot_theta}
\title{Plot priors and posteriors}
\usage{
plot_theta(ssm, path = NULL, id = NULL, x_limits = NULL)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{path}{character, where to find \code{X_*.csv}. If \code{NULL} (default), use the \code{path} of the last block (e.g. \code{/pmcmc}).}

\item{id}{numeric, indicate which \code{X_*.csv} to choose. If \code{NULL} (default), use the \code{id} of the last block (default to 0 in SSM).}
}
\value{
a \code{ggplot} object
}
\description{
Plot priors and posteriors distribution for all estimated parameters.
}
