% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_X}
\alias{plot_X}
\title{Plot states}
\usage{
plot_X(ssm, path = NULL, id = NULL, stat = c("none", "median", "mean"),
  hat = NULL, scales = "free_y", fit_only = FALSE, ran_only = FALSE,
  collapse_erlang = TRUE)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{path}{character, where to find \code{X_*.csv}. If \code{NULL} (default), use the \code{path} of the last block (e.g. \code{/pmcmc}).}

\item{id}{numeric, indicate which \code{X_*.csv} to choose. If \code{NULL} (default), use the \code{id} of the last block (default to 0 in SSM).}

\item{stat}{character, whether to plot a summary statistics of the state. Either \code{"mean"} or \code{"median"}. Default to \code{"none"}.}

\item{hat}{numeric, vector of credible intervals, between 0 and 1, e.g. \code{hat=c(0.5, 0.95)} for 50 and 95\% credible intervals.}

\item{scales}{character, should scales be \code{"fixed"}, \code{"free"}, or free in one dimension: \code{"free_x"}, \code{"free_y"} (the default).}

\item{fit_only}{logical, whether to show only the fit to the data.}

\item{ran_only}{logical, whether to show only the randomly generated data.}
}
\value{
a \code{ssm} object updated with latest SSM output and ready to be piped into another SSM block.
}
\description{
Function to plot state trajectories contained in the \code{X_*.csv} files generated by SSM. Pipeable.
}
\seealso{
\code{\link{plot_theta}}
}
