% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrap.r
\name{pmcmc}
\alias{pmcmc}
\title{Run a Particle-MCMC}
\usage{
pmcmc(ssm, approx = c("ode", "sde", "psr"), dt = NULL, id = 0,
  root = NULL, iter = NULL, n_parts = NULL, n_thread = 1,
  n_obs = NULL, cooling = NULL, switch = NULL, eps_switch = NULL,
  eps_max = NULL, eps_abs_integ = NULL, eps_rel_integ = NULL,
  smooth = FALSE, alpha = NULL, like_min = NULL, freeze_forcing = NULL,
  interpolator = NULL, verbose = FALSE, warning = FALSE,
  no_dem_sto = FALSE, no_white_noise = FALSE, no_diff = FALSE,
  traj = TRUE, n_traj = NULL, hat = FALSE, trace = TRUE, acc = TRUE,
  seed_time = TRUE)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{approx}{character, approximation used to simulate \code{ssm}:
\itemize{
    \item \code{"ode"} deterministic approximation, based on Ordinary Differential Equations.
    \item \code{"sde"} diffusion approximation, based on Stochastic Differential Equations.
    \item \code{"psr"} Euler-multinomial approximation, based on Poisson process with Stochastic Rates.
}}

\item{dt}{numeric, integration time step.}

\item{id}{integer, unique integer identifier that will be appended to the output.}

\item{root}{character, root path for output files (if any) (no trailing slash). If \code{NULL} (default), outputs are written in "your_model_path/the_name_of_the_wrapper".}

\item{iter}{numeric, number of iterations.}

\item{n_parts}{numeric, number of particles.}

\item{n_thread}{numeric, number of threads to be used. Default to 1. Use \code{"max"} to set it automatically to the number of cores available on the machine using \code{\link[parallel]{detectCores}}.}

\item{n_obs}{numeric, number of observations to be fitted (for tempering). If \code{NULL} (default), all observations are fitted.}

\item{cooling}{numeric, cooling factor (for sampling covariance live tuning or \code{mif} cooling).}

\item{switch}{numeric, select switching iteration from initial covariance to empirical one (\code{kmcmc} and \code{pmcmc}) or to update formula introduced in Ionides et al. 2006 (\code{mif})}

\item{eps_switch}{numeric, select number of burn-in iterations before tuning epsilon.}

\item{eps_max}{numeric, maximum value allowed for epsilon.}

\item{eps_abs_integ}{numeric, absolute error for adaptive step-size control.}

\item{eps_rel_integ}{numeric, relative error for adaptive step-size control.}

\item{smooth}{logical, tune epsilon with the value of the acceptance rate obtained with exponential smoothing. Default to \code{FALSE}.}

\item{alpha}{numeric, smoothing factor of exponential smoothing used to compute smoothed acceptance rate (low values increase degree of smoothing)}

\item{like_min}{numeric, particles with likelihood smaller than \code{like_min} are considered lost. If \code{NULL} (default) lower bound on likelihood based on machine precision.}

\item{freeze_forcing}{character, freeze covariates to their value at specified date (in \code{YYYY-MM-DD} format).}

\item{interpolator}{character, gsl interpolator for covariates}

\item{verbose}{logical, print logs (verbose). Default to \code{FALSE}.}

\item{warning}{logical, print warnings. Default to \code{FALSE}.}

\item{no_dem_sto}{logical, turn off demographic stochasticity (if any). Default to \code{FALSE}.}

\item{no_white_noise}{logical, turn off white noises (if any). Default to \code{FALSE}.}

\item{no_diff}{logical, turn off diffusions (if any). Default to \code{FALSE}.}

\item{traj}{logical, print the trajectories. Default to \code{TRUE}.}

\item{n_traj}{numeric, number of trajectories stored. If \code{NULL} (default), one trajectory is stored at each iteration.}

\item{hat}{logical, print the state estimates. Default to \code{FALSE}.}

\item{trace}{logical, print the trace. Default to \code{TRUE}.}

\item{acc}{logical, print the acceptance rate. Default to \code{TRUE}.}

\item{seed_time}{logical, seed the random number generator with the current time. Default to \code{TRUE}.}
}
\value{
a \code{ssm} object updated with latest SSM output and ready to be piped into another SSM block.
}
\description{
Function to run a MCMC using a SMC (particle filter) to evaluate the log-likelihood.
}

