% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.ssm}
\alias{print.ssm}
\title{Print ssm}
\usage{
\method{print}{ssm}(x, ...)
}
\arguments{
\item{x}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}
}
\description{
Function to print basic information on the \code{ssm}
}
