% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.r
\name{prior}
\alias{prior}
\alias{unif}
\alias{truncnorm}
\alias{dirac}
\title{Some priors distributions}
\usage{
unif(min, max)

truncnorm(a = -Inf, b = Inf, mean, sd)

dirac(value)
}
\arguments{
\item{min}{lower and upper limits of the distribution.  Must be finite.}

\item{max}{lower and upper limits of the distribution.  Must be finite.}

\item{a}{vector of lower bounds. These may be \code{-Inf}}

\item{b}{vector of upper bounds. These may be \code{Inf}}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{value}{numeric, dirac value}
}
\description{
Distributions supported by SSM to define priors
}
\seealso{
\code{\link{one_theta_sample_prior}}
}
