% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r2ssm.r
\name{r2ssm}
\alias{r2ssm}
\alias{r2ssm_prior}
\title{Convert R object into SSM object}
\usage{
r2ssm_prior(prior)
}
\arguments{
\item{prior}{a list}
}
\description{
Internal functions to convert R object into SSM
}

