% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prior.r
\name{sample_from_prior}
\alias{sample_from_prior}
\title{Sample from prior}
\usage{
sample_from_prior(priors, theta_names = NULL)
}
\arguments{
\item{priors}{a list of priors}

\item{theta_names}{character, names of the parameters sampled. If \code{NULL} (default) all parameters are sampled.}
}
\description{
Generate a sample from prior density distribution.
}

