% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prior.r
\name{sample_prior}
\alias{sample_prior}
\title{Sample prior}
\usage{
sample_prior(ssm, n, method = c("random"))
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{n}{numeric, sample size}

\item{method}{character, method used to sample from prior distribution:
\itemize{
    \item \code{"random"} randomly sample from the prior distribution
}}
}
\value{
if \code{n==1}: a \code{ssm} object, otherwise a \code{\link{tbl}} object with two columns: \code{id} (sampel index) and \code{ssm} (corresponding \code{ssm} object)
}
\description{
Generate one or more \code{ssm} objects with \code{theta} sampled from the prior distribution.
}
\seealso{
\code{\link{one_theta_sample_prior}} \code{\link{one_ssm_sample_prior}}
}

