% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.r
\name{sde}
\alias{diffusion}
\alias{sde}
\title{Define SDE on SSM inputs}
\usage{
diffusion(volatility, transformation = c("none", "log"))
}
\arguments{
\item{volatility}{character, name of the volatility parameter}

\item{transformation}{character, define a transformation for the input parameter (none by default). The diffusion is then defined for the transformed parameter:
\itemize{
        \item \code{"log"} log-transform for positive parameters.
}}
}
\description{
In SSM one can define stochastic differential equations on input parameters.
}
\examples{
# define a diffusion on log(beta) with volatility parameter "vol"	
input(name="beta", description="effective contact rate",transformation="R0/d_infectious", sde=diffusion(volatility="vol", transformation="log"))
}

