% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.r
\name{simul}
\alias{simul}
\title{Simulate ssm}
\usage{
simul(
  ssm,
  approx = c("ode", "sde", "psr"),
  dt = NULL,
  id = 0,
  root = NULL,
  n_thread = 1,
  n_parts = NULL,
  start = NULL,
  end = NULL,
  eps_abs_integ = NULL,
  eps_rel_integ = NULL,
  freeze_forcing = NULL,
  freq = NULL,
  interpolator = NULL,
  verbose = FALSE,
  warning = FALSE,
  no_dem_sto = FALSE,
  no_white_noise = FALSE,
  no_diff = FALSE,
  traj = TRUE,
  hat = FALSE,
  seed_time = TRUE
)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}

\item{approx}{character, approximation used to simulate \code{ssm}:
\itemize{
    \item \code{"ode"} deterministic approximation, based on Ordinary Differential Equations.
    \item \code{"sde"} diffusion approximation, based on Stochastic Differential Equations. 
    \item \code{"psr"} Euler-multinomial approximation, based on Poisson process with Stochastic Rates. 
}}

\item{dt}{numeric, integration time step.}

\item{id}{integer, unique integer identifier that will be appended to the output.}

\item{root}{character, root path for output files (if any) (no trailing slash). If \code{NULL} (default), outputs are written in "your_model_path/the_name_of_the_wrapper".}

\item{n_thread}{numeric, number of threads to be used. Default to 1. Use \code{"max"} to set it automatically to the number of cores available on the machine using \code{\link[parallel]{detectCores}}.}

\item{n_parts}{numeric, number of particles.}

\item{start, end}{character, starting and ending date of model simulation (in \code{YYYY-MM-DD} format). Default to data range.}

\item{eps_abs_integ}{numeric, absolute error for adaptive step-size control.}

\item{eps_rel_integ}{numeric, relative error for adaptive step-size control.}

\item{freeze_forcing}{character, freeze covariates to their value at specified date (in \code{YYYY-MM-DD} format).}

\item{freq}{numeric, for simulations outside the data range, print the outputs (and reset incidences to 0 if any) every specified days.}

\item{interpolator}{character, gsl interpolator for covariates}

\item{verbose}{logical, print logs (verbose). Default to \code{FALSE}.}

\item{warning}{logical, print warnings. Default to \code{FALSE}.}

\item{no_dem_sto}{logical, turn off demographic stochasticity (if any). Default to \code{FALSE}.}

\item{no_white_noise}{logical, turn off white noises (if any). Default to \code{FALSE}.}

\item{no_diff}{logical, turn off diffusions (if any). Default to \code{FALSE}.}

\item{traj}{logical, print the trajectories. Default to \code{TRUE}.}

\item{hat}{logical, print the state estimates. Default to \code{FALSE}.}

\item{seed_time}{logical, seed the random number generator with the current time. Default to \code{TRUE}.}
}
\value{
a \code{ssm} object updated with latest SSM output and ready to be piped into another SSM block.
}
\description{
Function to simulate a \code{ssm}.
}
