% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.r
\name{sympy_simplify}
\alias{sympy_simplify}
\title{Simplify symbolic expression}
\usage{
sympy_simplify(x, var_names)
}
\arguments{
\item{x}{character vector containing symbolic expressions to be simplified}

\item{var_names}{variables appearing in \code{x}}
}
\description{
Simplify symbolic epxression using rSymPy
}
