% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{to_tracer}
\alias{to_tracer}
\title{Export to tracer}
\usage{
to_tracer(ssm, path = NULL, id = NULL)
}
\arguments{
\item{ssm}{a \code{ssm} object, returned by \code{\link{new_ssm}}.}
}
\description{
Export to tracer
}

