% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input.r
\name{input}
\alias{input}
\title{Create SSM input}
\usage{
input(name, description = NULL, value = NULL, prior = NULL,
  transformation = NULL, sde = NULL, tag = c("none", "remainder",
  "pop_size"))
}
\arguments{
\item{name}{character, name of the input}

\item{description}{character, description of the input}

\item{value}{numeric or character, define the value of the input}

\item{prior}{define the prior of the input, as returned by a \code{\link{prior}} helper}

\item{transformation}{define the transformation of the input (see example)}

\item{sde}{define a stochastic differential equation on the input (see example)}

\item{tag}{character, tag for specific inputs. Set to one among:}
}
\description{
Create an input parameter for SSM
}
\examples{
\dontrun{
 TODO
}
}
\seealso{
\code{\link{prior}}
}

