% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/observation.r
\name{obs}
\alias{discretized_normal_obs}
\alias{obs}
\alias{poisson_obs}
\title{Observation processes for SSM}
\usage{
poisson_obs(obs = NULL, state = NULL, reporting)

discretized_normal_obs(obs = NULL, state = NULL, reporting, overdispersion)
}
\arguments{
\item{obs}{character, name of the observation in the data}

\item{state}{character, name of the observed state in the model}

\item{reporting}{character, name of the reporting rate parameter}

\item{overdispersion}{character, name of the overdispersion parameter}
}
\description{
SSM currently accept the following observation processes.
}

