% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2ssm.r
\name{r2ssm}
\alias{r2ssm}
\alias{r2ssm_prior}
\alias{r2ssm}
\alias{r2ssm_resources}
\alias{r2ssm}
\alias{ssm2r_resources}
\alias{r2ssm}
\alias{r2ssm_populations}
\title{Convert R object into SSM object}
\usage{
r2ssm_prior(prior)

r2ssm_resources(theta, covmat)

ssm2r_resources(ssm_theta)

r2ssm_populations(pop, state_variables, remainder = NULL, pop_size = NULL)
}
\arguments{
\item{prior}{a list}

\item{theta}{a vector}

\item{covmat}{a matrix}

\item{ssm_theta}{a list (usually a theta.json of SSM parsed by \code{\link[rjson]{fromJSON}})}

\item{pop}{character}

\item{state_variables}{character vector, name of all state variables}

\item{remainder}{character, name of the state variable considered as a remainder.}

\item{pop_size}{character, name of the parameter corresponding to the population size. Required if \code{remainder} is provided.}
}
\description{
Internal functions to convert R object into SSM
}
