% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reaction.r
\name{reaction}
\alias{reaction}
\title{Create SSM reaction}
\usage{
reaction(from, to, description = NULL, rate, accumulators = NULL)
}
\arguments{
\item{from,to}{character, name of starting and ending state of the reaction}

\item{description}{character, description of the reaction}

\item{rate}{character, rate of the reaction}

\item{accumulators}{character vector, one or more accumulators}
}
\description{
Create a reaction for SSM
}

